feature: create hamburger menu component and implement absolute and inde import/exports
