+++ components/HamburgerMenu/index.ts
@@ -1,0 +1,3 @@
+import { HamburgerMenu } from "./HamburgerMenu";
+
+export { HamburgerMenu };
