+++ components/LandingPage/index.tsx
@@ -1,0 +1,3 @@
+import { LandingPage } from "./LandingPage";
+
+export { LandingPage };
