+++ pages/developer/index.tsx
@@ -1,0 +1,5 @@
+export const Developer = () => (
+  <div>
+    <p>Hello Developer</p>
+  </div>
+);
