+++ pages/index.tsx
@@ -1,5 +1,5 @@
-import { LandingPage } from '../components/LandingPage'
+import { LandingPage } from "@components/LandingPage";
 
-export default function Home() {
-  <LandingPage/>
-}
+const Home = () => <LandingPage />;
+
+export default Home;
