+++ pages/projects/index.tsx
@@ -1,0 +1,5 @@
+export const Projects = () => (
+    <div>
+        <p>Hello Projects</p>
+    </div>
+)
