+++ styles/HamburgerMenu.module.css
@@ -1,0 +1,35 @@
+body {
+    padding: 0px;
+    margin: 0px;
+    font-size: 20px;
+    min-height: 100vh;
+  }
+  
+  #hamburger-menu {
+      width: 50px;
+      height: 50px;
+      display: block;
+      border: none;
+      padding: 0px;
+      margin: 20px;
+      background: linear-gradient(
+        to bottom, 
+        #3D0E61, #3D0E61 20%, 
+        white 20%, white 40%, 
+        #3D0E61 40%, #3D0E61 60%, 
+        white 60%, white 80%, 
+        #3D0E61 80%, #3D0E61 100%
+      );
+  }
+  
+  #hamburger-menu #sidebar-menu {
+      visibilty: hidden;
+      position: fixed;
+      top: 0;
+      left: -250px;
+      width: 200px;
+      height: 100%;
+      background-color: #3D0E61;
+      padding: 0px 10px;
+      box-sizing: border-box;
+  }
